+++ src/common/VersionedSchema/Utils/__mocks__/compatibilityCases.ts
@@ -114,7 +114,7 @@ export const compatibilityCases: CompatibilityCase[] = [
             { type: 'added', key: 'models' },
             { type: 'added', key: 'analyze' },
             { type: 'added', key: 'miracles' },
-        ],
+        ]
     },
     {
         type: EVersionedSchemaType.UNIFIED_OPTIONS,
@@ -123,6 +123,8 @@ export const compatibilityCases: CompatibilityCase[] = [
         expect: [
             { type: 'added', key: 'strictOpenapi' },
             { type: 'added', key: 'reportFile' },
+            { type: 'added', key: 'useProjectPrettier' },
+            { type: 'added', key: 'useEslintFix' },
         ],
     },
 ];